+++ plugins/example/src/index.ts
@@ -1,0 +1,6 @@
+export type { ExampleSessionView, ExampleSnapshot } from "./types.js";
+export { EXAMPLE_PROFILE, resolveExampleTopics } from "./profile.js";
+export { buildExampleSnapshot } from "./build-snapshot.js";
+export { EXAMPLE_PLUGIN_MANIFEST } from "./manifest.js";
+export { ExampleDock } from "./ExampleDock.js";
+export { useExampleSnapshot } from "./hooks/useExampleSnapshot.js";